Move the CLI to uvicorn/__main__.py and keep run() in uvicorn/main.py
